Cleanup existing properties and make the pattern consistent
